src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.2.1
    # via httpcore
arrow==1.1.1
    # via -r requirements.in
authlib==0.15.4
    # via tda-api
autopep8==1.5.7
    # via tda-api
boto3==1.17.110
    # via -r requirements.in
boto3-stubs[s3]==1.17.112
    # via -r requirements.in
botocore==1.20.110
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.20.112
    # via boto3-stubs
certifi==2021.5.30
    # via httpx
cffi==1.14.6
    # via cryptography
cryptography==3.4.7
    # via authlib
h11==0.12.0
    # via httpcore
httpcore==0.13.6
    # via httpx
httpx==0.18.1
    # via tda-api
idna==3.2
    # via
    #   anyio
    #   rfc3986
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
mypy-boto3-s3==1.17.112
    # via boto3-stubs
numpy==1.21.0
    # via pandas
pandas==1.3.0
    # via -r requirements.in
prompt-toolkit==3.0.19
    # via tda-api
pycodestyle==2.7.0
    # via autopep8
pycparser==2.20
    # via cffi
python-dateutil==2.8.1
    # via
    #   arrow
    #   botocore
    #   pandas
    #   tda-api
pytz==2021.1
    # via pandas
rfc3986[idna2008]==1.5.0
    # via httpx
s3transfer==0.4.2
    # via boto3
selenium==3.141.0
    # via tda-api
six==1.16.0
    # via python-dateutil
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tda-api==1.3.4
    # via -r requirements.in
tenacity==8.0.1
    # via -r requirements.in
toml==0.10.2
    # via autopep8
urllib3==1.26.6
    # via
    #   botocore
    #   selenium
wcwidth==0.2.5
    # via prompt-toolkit
websockets==9.1
    # via tda-api
